pyproject.toml
[tool.poetry]
name = "job-manager"
version = "0.1.0"
description = "Manage jobs in the paradim platform."
authors = ["Gabriel Couture <gacou54@ulaval.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = ">=4.2,<5"
djangorestframework = "^3.15.1"
whitenoise = "^6.6.0"
psycopg2 = "^2.9.9"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
celery = "^5.4.0"
python-dotenv = "^1.0.1"
django-jazzmin = "^3.0.0"
parameterized = "^0.9.0"
httpx = "^0.27.0"
pyorthanc = "^1.16.1"
paramiko = "^3.4.0"
gunicorn = "^22.0.0"
msal = "^1.29.0"
pymemcache = "^4.0.0"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
